+++ crates/ruma-client-api/src/lib.rs
@@ -43,6 +43,8 @@ pub mod relations;
 pub mod rendezvous;
 pub mod reporting;
 pub mod room;
+#[cfg(feature = "unstable-msc4143")]
+pub mod rtc;
 pub mod search;
 pub mod server;
 pub mod session;