+++ crates/ruma-client-api/src/rtc.rs
@@ -1,0 +1,5 @@
+//! [MatrixRTC] endpoints.
+//!
+//! [MatrixRTC]: https://github.com/matrix-org/matrix-spec-proposals/pull/4143
+
+pub mod transports;